refactor(utils): extract profiling helpers to dedicated module
- Move cprofile_context and cprofile from vllm.utils.__init__ to
  vllm/utils/profiling.py
- Implement PEP 562 module-level __getattr__ to emit DeprecationWarning
  for old import paths (vllm.utils.cprofile -> vllm.utils.profiling.cprofile)
- Add __dir__ for proper tab-completion support
- Remove silent re-export to make deprecation explicit and guide users
  to the new API

This refactoring improves code organization and provides clear migration
path for users through proper deprecation warnings.
